requirements.txt
# PDF Table Extractor with Advanced Image Processing and Line Detection
# Core dependencies for image processing and OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0

# Data processing and export
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Web interface (optional)
streamlit>=1.28.0

# PDF processing
PyMuPDF>=1.23.0
